Create Financial-Educational-Game.py
There's a certain degree of luck, the roll of the dice, and a certain degree of skill. It's like poker. And, unlike many games, Monopoly has built-in flexibility. Players can make up their own rules - paying money for landing on free parking, for example - without "unbalancing" the game.

Rigged games are unfair and no fun, and our goal as game developers is always to deliver the best gaming experience possible.